requirements.txt
annotated-types==0.7.0
anyio==4.3.0
assemblyai==0.26.0
blinker==1.8.2
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
rev_ai==2.19.4
pytz==2024.1
Flask-Login==0.6.3
websocket-client==0.59.0
distro==1.9.0
openai==1.35.7
tqdm==4.66.4
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
tomli==2.0.1
pynamodb==6.0.1
